app/requirements.txt
flask
uwsgi
requests
mysql-connector-python-rf
